Use a constant ECOS_VERSION instead of recomputing version
This moves the ECOS version from being recomputed via a `ccall` into the
library on demand to a compile-time constant.

It also adds a missing call to `check_deps` in `__init__`, which is
recommended by all applications using BinaryProvider, and removes the
note about Homebrew from the error message, as that is no longer
relevant.